backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
pydantic-settings==2.7.1
python-multipart==0.0.20
psycopg[binary]==3.2.13
supabase==2.28.3
pyjwt==2.12.1